requirements.txt
bcrypt
eventlet
flask
flask-socketio
flask-sqlalchemy
kafka-python
marshmallow
python-dotenv
pyjwt
passlib[bcrypt]
psycopg2-binary
requests
sqlalchemy